Do not mess with MDADMs auto-detect
Only reason for this logic is to limit what entries in /proc/partitions
mdadm scans for arrays.
However this seems to break raids in containers (example Intel fakeraids)
as those containers are never scanned for arrays.
(On my system it activates /dev/md/imsm0, but never /dev/md/Gentoo_0 which
is the real RAID with the removed logic)
If someone has something dead-slow they do not want scanned for raids,
it is better if they specify their own mdadm.conf.
If someone finds an example of something that mdadm scans for arrays that
is not supposed to be scanned, we may handle that then.