Support for glibc-2.14 systems
Hello,

Due to the changes in glibc-2.14 busybox no longer compiles when NFS
support is enabled.  This breaks genkernel.  The following is a very
crude patch to just make things work on glibc-2.14 systems.  A better
patch would be to detect glibc-2.14 and then make a config decision.
The best patch would be to port libtirpc to openssl, make busybox link
against libtirpc, restore NFS support, and put an end to this whole
mess.  But in the meantime:

commit 80d1e89995f055a37f07f9f7f4c7e21660abd527
Author: Borg Onion <borg.onion@gmail.com>
Date:   Sat Dec 3 18:02:48 2011 -0800

    Disable NFS support to enable glibc-2.14 compatibility