gen_initramfs: Release the copy_binaries function as public domain.
The copy_binaries function is explicitly released into the public domain to
encourage wide adoption and re-use.

This is to enable use in some cases of BSD and other licensing, and
specifically avoiding the need to be party to any contributor agreements.
Additionally, the code could be considered of trivial length in some regions.

Written by:
- Robin H. Johnson <robbat2@gentoo.org> (complete rewrite)
- Richard Yao <ryao@cs.stonybrook.edu> (original concept)

Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>