Remove duplicate files from initramfs
genkernel's initramfs image is built incrementally by appending to the
cpio file. The introduction of copy_binaries resulted in copying
libraries from the host system, which causes the cpio to include certain
libraries multiple times whenever different stages depended upon the
same library. We address this by extracting the cpio to a temporary
directory and then compressing it again to "finalize" it. The extraction
eliminates the duplicate files.

This makes generated initramfs images slightly smaller and in theory,
should make the initramfs load slightly faster.

Signed-off-by: Richard Yao <ryao@gentoo.org>

Conflicts:
	gen_initramfs.sh