00-rootdev.sh: ensure_initramfs_mounts(): fsck fs before mount
This change adds support for checking the integrity of
filesystems mounted from within the initramfs.

Systemd wants the filesystem on which /usr resides mounted when
it finally takes over from initramfs, but does not seem to
launch a systemd-fsck job for that filesystem later on, even if
the filesystem is mounted ro.  The result is that the
/usr-hosting filesystem could go unchecked for a long time.

The problem is not limited to /usr, but all early mounts
mentioned in initramfs.mounts.

Furthermore, fsck'ing should work on the physical device, final
mounting should take into account existing symbolic or logical
disk/LVM names for the purpose of enhanced readability.