Use KERNEL_OUTPUTDIR for integrated initramfs configuration
--integrated-initramfs builds the kernel in 2 passes, and modifies the kernel configuration in-between.
However it used ${KERNEL_DIR}/.config, which is in the kernel source tree. While this did work for in-tree builds, it made out-of-tree builds fail since the source tree then appeared unclean (and the written .config file only contained the CONFIG_INITRAMFS_* items anyway).
This commit switches to ${KERNEL_OUTPUTDIR}/.config which is the right path both for in-tree and out-of-tree builds.

Also, some additional quotes are added.